+++ crates/core/src/repository/backend/mod.rs
@@ -6,13 +6,13 @@ use dyn_clone::{clone_trait_object, DynClone};
 
 pub mod curse;
 pub mod git;
-pub mod townlongyak;
+pub mod hub;
 pub mod tukui;
 pub mod wowi;
 
 pub use curse::Curse;
 pub use git::{Github, Gitlab};
-pub use townlongyak::TownlongYak;
+pub use hub::Hub;
 pub use tukui::Tukui;
 pub use wowi::WowI;