+++ CHANGELOG.md
@@ -14,6 +14,20 @@ and `Removed`.
 
 ## [Unreleased]
 
+### Added
+
+- Possible to see all addons from all sources when browsing the Catalog by selecting
+  "All Sources" in the list.
+- Hub catalog source.
+  - This is a new source which the [WowUp](https://wowup.io/) team has created.
+
+### Removed
+
+- Townlong Yak source has been removed from Catalog due to it being [deprecated](https://www.townlong-yak.com/addons/about/update-api)
+  by the author.
+  - Townlong Yak addons are still available through both CurseForge and the newly
+    added source, Hub.
+
 ### Changed
 
 - Speed up parsing addons by improving regex expessions.